refactor: extract Swagger docs and split BffAdminService for cleaner architecture
- Extract all Swagger documentation into separate bff-admin.swagger.ts file
- Split large BffAdminService (934 lines) into domain-specific services:
  * BffAdminTeacherService - handles teacher-related operations
  * BffAdminStudentService - handles student-related operations
  * BffAdminClassroomService - handles classroom-related operations
- Update main BffAdminService to act as orchestrator (60 lines)
- Clean up controller by using extracted Swagger decorators
- Maintain same API while improving code organization and maintainability
- Follow Single Responsibility Principle for better separation of concerns

Files changed:
- src/components/bff/admin/bff-admin.service.ts (refactored to orchestrator)
- src/components/bff/admin/bff-admin.controller.ts (cleaned up with extracted decorators)
- src/components/bff/admin/bff-admin.swagger.ts (new file for all Swagger docs)
- src/components/bff/admin/services/ (new directory with domain services)
- src/components/bff/admin/bff-admin.module.ts (updated providers)

Benefits:
- Improved maintainability and readability
- Better separation of concerns
- Easier testing and debugging
- Reduced file sizes for better navigation
- Cleaner controller architecture
