+++ src/components/bff/admin/bff-admin.service.ts
@@ -1,92 +1,29 @@
 import { Injectable } from '@nestjs/common';
-import { TeacherStatus, EmploymentType } from '@prisma/client';
 
 import { PrismaService } from '../../../prisma';
+import { BffAdminClassroomService } from './services/bff-admin-classroom.service';
+import { BffAdminStudentService } from './services/bff-admin-student.service';
+import { BffAdminTeacherService } from './services/bff-admin-teacher.service';
 import {
   AdminClassroomsViewData,
   ClassroomDetailsData,
-  PaginatedStudents,
   PaginatedStudentDetails,
   SingleStudentDetails,
-  TeachersViewData,
   SingleTeacherDetails,
+  TeachersViewData,
 } from './types';
 
 @Injectable()
 export class BffAdminService {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(
+    private readonly prisma: PrismaService,
+    private readonly teacherService: BffAdminTeacherService,
+    private readonly studentService: BffAdminStudentService,
+    private readonly classroomService: BffAdminClassroomService,
+  ) {}
 
   async getClassroomsViewData(userId: string): Promise<AdminClassroomsViewData> {
-    // First, get the user's school ID
-    const user = await this.prisma.user.findUnique({
-      where: { id: userId },
-      select: { schoolId: true },
-    });
-
-    if (!user?.schoolId) {
-      throw new Error('User not found or not associated with a school');
-    }
-
-    const schoolId = user.schoolId;
-
-    // Get all class arms for the school with their students and teachers
-    const classArms = await this.prisma.classArm.findMany({
-      where: { schoolId },
-      include: {
-        level: true,
-        students: true,
-        classTeacher: {
-          include: {
-            user: true,
-          },
-        },
-        captain: {
-          include: {
-            user: true,
-          },
-        },
-      } as any, // Temporary until Prisma types fully refresh
-    });
-
-    // Calculate stats (with type assertions until Prisma types fully refresh)
-    const totalClassrooms = classArms.length;
-    const totalStudents = (classArms as any[]).reduce(
-      (sum, classArm) => sum + classArm.students.length,
-      0,
-    );
-
-    // Get unique grade levels
-    const gradeLevels = new Set((classArms as any[]).map((ca) => ca.level.name)).size;
-
-    // Calculate capacity usage (placeholder calculation)
-    const capacityUsage = Math.min(100, (totalStudents / (totalClassrooms * 30)) * 100);
-
-    // Prepare classroom list using efficient direct references
-    const classroomList = classArms.map((classArm) => {
-      return {
-        id: classArm.id,
-        name: classArm.name,
-        level: (classArm as any).level.name,
-        location: null, // Add location field to ClassArm model if needed
-        classTeacher: (classArm as any).classTeacher
-          ? `${(classArm as any).classTeacher.user.firstName} ${(classArm as any).classTeacher.user.lastName}`
-          : null,
-        classCaptain: (classArm as any).captain
-          ? `${(classArm as any).captain.user.firstName} ${(classArm as any).captain.user.lastName}`
-          : null,
-        studentsCount: (classArm as any).students.length,
-      };
-    });
-
-    return {
-      stats: {
-        totalClassrooms,
-        totalStudents,
-        gradeLevels,
-        capacityUsage: Math.round(capacityUsage * 100) / 100,
-      },
-      classrooms: classroomList,
-    };
+    return this.classroomService.getClassroomsViewData(userId);
   }
 
   async getClassroomDetailsData(
@@ -95,266 +32,7 @@ export class BffAdminService {
     page: number = 1,
     limit: number = 20,
   ): Promise<ClassroomDetailsData> {
-    // First, get the user's school ID
-    const user = await this.prisma.user.findUnique({
-      where: { id: userId },
-      select: { schoolId: true },
-    });
-
-    if (!user?.schoolId) {
-      throw new Error('User not found or not associated with a school');
-    }
-
-    const schoolId = user.schoolId;
-
-    // Get classroom details with efficient direct references
-    const classroom = await this.prisma.classArm.findFirst({
-      where: {
-        id: classroomId,
-        schoolId,
-      },
-      include: {
-        level: true,
-        classTeacher: {
-          include: {
-            user: true,
-          },
-        },
-        captain: {
-          include: {
-            user: true,
-          },
-        },
-      } as any, // Type assertion until VS Code restarts
-    });
-
-    if (!classroom) {
-      throw new Error('Classroom not found or not accessible');
-    }
-
-    // Get total count of students in the classroom
-    const totalStudents = await this.prisma.student.count({
-      where: {
-        classArmId: classroomId,
-      },
-    });
-
-    // Calculate pagination
-    const skip = (page - 1) * limit;
-    const totalPages = Math.ceil(totalStudents / limit);
-
-    // Get paginated students with all their data
-    const paginatedStudents = await this.prisma.student.findMany({
-      where: {
-        classArmId: classroomId,
-      },
-      skip,
-      take: limit,
-      include: {
-        user: {
-          include: {
-            address: true, // Include address for residential info
-          },
-        },
-        prefect: true,
-        guardian: {
-          include: {
-            user: true, // Include guardian's user data to get their name
-          },
-        },
-        subjectTermStudents: {
-          take: 50, // Limit to recent results
-          orderBy: {
-            createdAt: 'desc',
-          },
-          include: {
-            subjectTerm: {
-              include: {
-                subject: true,
-              },
-            },
-            assessments: {
-              orderBy: {
-                createdAt: 'desc',
-              },
-            },
-          },
-        },
-      },
-    });
-
-    // Get all students for population and top performers calculation (without pagination)
-    const allStudents = await this.prisma.student.findMany({
-      where: {
-        classArmId: classroomId,
-      },
-      include: {
-        user: {
-          include: {
-            address: true, // Include address for residential info
-          },
-        },
-        prefect: true,
-        guardian: {
-          include: {
-            user: true, // Include guardian's user data
-          },
-        },
-        subjectTermStudents: {
-          take: 50,
-          orderBy: {
-            createdAt: 'desc',
-          },
-          include: {
-            subjectTerm: {
-              include: {
-                subject: true,
-              },
-            },
-            assessments: {
-              orderBy: {
-                createdAt: 'desc',
-              },
-            },
-          },
-        },
-      },
-    });
-
-    // Calculate population statistics
-    const maleStudents = allStudents.filter((student) => student.user.gender === 'MALE').length;
-    const femaleStudents = allStudents.filter((student) => student.user.gender === 'FEMALE').length;
-
-    // Calculate attendance statistics
-    const totalDays = 180; // Academic year days
-    const presentDays = Math.floor(totalDays * (0.85 + Math.random() * 0.15)); // Random between 85-100%
-    const absentDays = totalDays - presentDays;
-    const attendanceRate = (presentDays / totalDays) * 100;
-
-    // Calculate individual student attendance for today (realistic simulation)
-    const studentsPresent = Math.floor(totalStudents * (0.9 + Math.random() * 0.08)); // 90-98% present today
-    const studentsAbsent = totalStudents - studentsPresent;
-
-    // Get class teacher and captain using efficient direct references
-    const classTeacher = (classroom as any).classTeacher;
-    const classCaptain = (classroom as any).captain;
-
-    // Calculate top performers from assessment results
-    const performanceMap = new Map<
-      string,
-      { totalScore: number; count: number; student: any; latestSubject: string }
-    >();
-
-    allStudents.forEach((student) => {
-      student.subjectTermStudents.forEach((subjectTermStudent) => {
-        const studentId = student.id;
-        const existing = performanceMap.get(studentId) || {
-          totalScore: 0,
-          count: 0,
-          student,
-          latestSubject: subjectTermStudent.subjectTerm.subject.name,
-        };
-
-        existing.totalScore += subjectTermStudent.totalScore;
-        existing.count += 1;
-        performanceMap.set(studentId, existing);
-      });
-    });
-
-    const topPerformers = Array.from(performanceMap.entries())
-      .map(([studentId, data]) => ({
-        id: studentId,
-        name: `${data.student.user.firstName} ${data.student.user.lastName}`,
-        score: data.count > 0 ? data.totalScore / data.count : 0,
-        subject: data.latestSubject,
-      }))
-      .sort((a, b) => b.score - a.score)
-      .slice(0, 3);
-
-    // Prepare paginated student information
-    const studentsData = paginatedStudents.map((student) => {
-      const age = student.user.dateOfBirth
-        ? Math.floor(
-            (Date.now() - new Date(student.user.dateOfBirth).getTime()) /
-              (365.25 * 24 * 60 * 60 * 1000),
-          )
-        : 0;
-
-      // Get guardian name
-      const guardianName = student.guardian
-        ? `${student.guardian.user.firstName} ${student.guardian.user.lastName}`
-        : 'N/A';
-
-      // Get state of origin from user's address
-      const stateOfOrigin = student.user.stateOfOrigin || 'N/A';
-
-      return {
-        id: student.id,
-        name: `${student.user.firstName} ${student.user.lastName}`,
-        gender: student.user.gender,
-        age,
-        admissionNumber: student.admissionNo || 'N/A',
-        guardianPhone: student.user.phone,
-        guardianName,
-        stateOfOrigin,
-      };
-    });
-
-    // Create pagination info
-    const paginationInfo = {
-      page,
-      limit,
-      total: totalStudents,
-      totalPages,
-      hasNext: page < totalPages,
-      hasPrevious: page > 1,
-    };
-
-    // Create paginated students response
-    const students: PaginatedStudents = {
-      students: studentsData,
-      pagination: paginationInfo,
-    };
-
-    return {
-      classroom: {
-        id: classroom.id,
-        name: classroom.name,
-        level: (classroom as any).level.name,
-        location: null, // Add location field to ClassArm model if needed
-      },
-      population: {
-        total: totalStudents,
-        male: maleStudents,
-        female: femaleStudents,
-      },
-      attendance: {
-        totalDays,
-        presentDays,
-        absentDays,
-        attendanceRate: Math.round(attendanceRate * 100) / 100,
-        studentsPresent,
-        studentsAbsent,
-        totalStudents,
-      },
-      classTeacher: classTeacher
-        ? {
-            id: classTeacher.id,
-            name: `${classTeacher.user.firstName} ${classTeacher.user.lastName}`,
-            phone: classTeacher.user.phone,
-            email: classTeacher.user.email,
-          }
-        : null,
-      classCaptain: classCaptain
-        ? {
-            id: classCaptain.id,
-            name: `${classCaptain.user.firstName} ${classCaptain.user.lastName}`,
-            admissionNumber: classCaptain.admissionNo,
-          }
-        : null,
-      students,
-      topPerformers,
-    };
+    return this.classroomService.getClassroomDetailsData(userId, classroomId, page, limit);
   }
 
   async getStudentDetailsData(
@@ -364,571 +42,18 @@ export class BffAdminService {
     classroomId?: string,
     search?: string,
   ): Promise<PaginatedStudentDetails> {
-    // First, get the user's school ID
-    const user = await this.prisma.user.findUnique({
-      where: { id: userId },
-      select: { schoolId: true },
-    });
-
-    if (!user?.schoolId) {
-      throw new Error('User not found or not associated with a school');
-    }
-
-    const schoolId = user.schoolId;
-
-    // Build where clause for filtering
-    const whereClause: any = {
-      user: {
-        schoolId,
-      },
-    };
-
-    // Add classroom filter if provided
-    if (classroomId) {
-      whereClause.classArmId = classroomId;
-    }
-
-    // Add search filter if provided
-    if (search) {
-      whereClause.OR = [
-        {
-          user: {
-            firstName: {
-              contains: search,
-              mode: 'insensitive',
-            },
-          },
-        },
-        {
-          user: {
-            lastName: {
-              contains: search,
-              mode: 'insensitive',
-            },
-          },
-        },
-        {
-          admissionNo: {
-            contains: search,
-            mode: 'insensitive',
-          },
-        },
-        {
-          studentNo: {
-            contains: search,
-            mode: 'insensitive',
-          },
-        },
-      ];
-    }
-
-    // Get total count
-    const totalStudents = await this.prisma.student.count({
-      where: whereClause,
-    });
-
-    // Calculate pagination
-    const skip = (page - 1) * limit;
-    const totalPages = Math.ceil(totalStudents / limit);
-
-    // Get paginated students with all necessary data
-    const students = await this.prisma.student.findMany({
-      where: whereClause,
-      skip,
-      take: limit,
-      include: {
-        user: {
-          include: {
-            address: true,
-          },
-        },
-        guardian: {
-          include: {
-            user: true,
-          },
-        },
-        classArm: {
-          include: {
-            level: true,
-          },
-        },
-        subjectTermStudents: {
-          include: {
-            assessments: true,
-          },
-        },
-        studentAttendances: {
-          orderBy: {
-            createdAt: 'desc',
-          },
-          take: 30, // Last 30 attendance records for rate calculation
-        },
-      },
-      orderBy: [
-        {
-          user: {
-            firstName: 'asc',
-          },
-        },
-        {
-          user: {
-            lastName: 'asc',
-          },
-        },
-      ],
-    });
-
-    // Process student data
-    const studentsData = students.map((student) => {
-      // Calculate age
-      const age = student.user.dateOfBirth
-        ? Math.floor(
-            (new Date().getTime() - new Date(student.user.dateOfBirth).getTime()) /
-              (365.25 * 24 * 60 * 60 * 1000),
-          )
-        : 0;
-
-      // Calculate average grade from assessments
-      const allScores = student.subjectTermStudents.flatMap((sts) =>
-        sts.assessments.map((assessment) => assessment.score),
-      );
-      const averageGrade =
-        allScores.length > 0
-          ? allScores.reduce((sum, score) => sum + score, 0) / allScores.length
-          : 0;
-
-      // Calculate attendance rate
-      const presentCount = student.studentAttendances.filter(
-        (attendance) => attendance.status === 'PRESENT',
-      ).length;
-      const totalAttendanceRecords = student.studentAttendances.length;
-      const attendanceRate =
-        totalAttendanceRecords > 0 ? (presentCount / totalAttendanceRecords) * 100 : 0;
-
-      // Simulate current presence (random for demo purposes)
-      const isPresent = Math.random() > 0.1; // 90% chance of being present
-
-      // Get guardian info
-      const guardianName = student.guardian
-        ? `${student.guardian.user.firstName} ${student.guardian.user.lastName}`
-        : 'N/A';
-
-      const guardianPhone = student.guardian?.user.phone || null;
-
-      return {
-        id: student.id,
-        name: `${student.user.firstName} ${student.user.lastName}`,
-        studentId: student.studentNo,
-        admissionNumber: student.admissionNo || 'N/A',
-        gender: student.user.gender === 'MALE' ? 'Male' : 'Female',
-        age,
-        stateOfOrigin: student.user.stateOfOrigin || 'N/A',
-        guardianName,
-        guardianPhone,
-        telephone: student.user.phone,
-        className: student.classArm ? student.classArm.name : 'N/A',
-        classLevel: student.classArm?.level ? student.classArm.level.name : 'N/A',
-        averageGrade: Math.round(averageGrade * 100) / 100,
-        isPresent,
-        attendanceRate: Math.round(attendanceRate * 100) / 100,
-        avatarUrl: student.user.avatarUrl,
-      };
-    });
-
-    // Create pagination info
-    const paginationInfo = {
-      page,
-      limit,
-      total: totalStudents,
-      totalPages,
-      hasNext: page < totalPages,
-      hasPrevious: page > 1,
-    };
-
-    return {
-      students: studentsData,
-      pagination: paginationInfo,
-    };
+    return this.studentService.getStudentDetailsData(userId, page, limit, classroomId, search);
   }
 
   async getSingleStudentDetails(userId: string, studentId: string): Promise<SingleStudentDetails> {
-    // First, get the user's school ID
-    const user = await this.prisma.user.findUnique({
-      where: { id: userId },
-      select: { schoolId: true },
-    });
-
-    if (!user?.schoolId) {
-      throw new Error('User not found or not associated with a school');
-    }
-
-    const schoolId = user.schoolId;
-
-    // Get the specific student with all necessary data
-    const student = await this.prisma.student.findFirst({
-      where: {
-        id: studentId,
-        user: {
-          schoolId,
-        },
-      },
-      include: {
-        user: {
-          include: {
-            address: true,
-          },
-        },
-        guardian: {
-          include: {
-            user: true,
-          },
-        },
-        classArm: {
-          include: {
-            level: true,
-          },
-        },
-        subjectTermStudents: {
-          include: {
-            assessments: true,
-          },
-        },
-        studentAttendances: {
-          orderBy: {
-            createdAt: 'desc',
-          },
-          take: 30, // Last 30 attendance records for rate calculation
-        },
-      },
-    });
-
-    if (!student) {
-      throw new Error('Student not found or not accessible');
-    }
-
-    // Calculate age
-    const age = student.user.dateOfBirth
-      ? Math.floor(
-          (new Date().getTime() - new Date(student.user.dateOfBirth).getTime()) /
-            (365.25 * 24 * 60 * 60 * 1000),
-        )
-      : 0;
-
-    // Calculate average grade from assessments
-    const allScores = student.subjectTermStudents.flatMap((sts) =>
-      sts.assessments.map((assessment) => assessment.score),
-    );
-    const averageGrade =
-      allScores.length > 0
-        ? allScores.reduce((sum, score) => sum + score, 0) / allScores.length
-        : 0;
-
-    // Calculate attendance rate
-    const presentCount = student.studentAttendances.filter(
-      (attendance) => attendance.status === 'PRESENT',
-    ).length;
-    const totalAttendanceRecords = student.studentAttendances.length;
-    const attendanceRate =
-      totalAttendanceRecords > 0 ? (presentCount / totalAttendanceRecords) * 100 : 0;
-
-    // Simulate current presence (random for demo purposes)
-    const isPresent = Math.random() > 0.1; // 90% chance of being present
-
-    // Get guardian info
-    const guardianName = student.guardian
-      ? `${student.guardian.user.firstName} ${student.guardian.user.lastName}`
-      : 'N/A';
-
-    const guardianPhone = student.guardian?.user.phone || null;
-
-    return {
-      id: student.id,
-      name: `${student.user.firstName} ${student.user.lastName}`,
-      studentId: student.studentNo,
-      admissionNumber: student.admissionNo || 'N/A',
-      gender: student.user.gender === 'MALE' ? 'Male' : 'Female',
-      age,
-      stateOfOrigin: student.user.stateOfOrigin || 'N/A',
-      guardianName,
-      guardianPhone,
-      telephone: student.user.phone,
-      className: student.classArm ? student.classArm.name : 'N/A',
-      classLevel: student.classArm?.level ? student.classArm.level.name : 'N/A',
-      averageGrade: Math.round(averageGrade * 100) / 100,
-      isPresent,
-      attendanceRate: Math.round(attendanceRate * 100) / 100,
-      avatarUrl: student.user.avatarUrl,
-    };
+    return this.studentService.getSingleStudentDetails(userId, studentId);
   }
 
   async getTeachersViewData(userId: string): Promise<TeachersViewData> {
-    // First, get the user's school ID
-    const user = await this.prisma.user.findUnique({
-      where: { id: userId },
-      select: { schoolId: true },
-    });
-
-    if (!user?.schoolId) {
-      throw new Error('User not found or not associated with a school');
-    }
-
-    const schoolId = user.schoolId;
-
-    // Get all teachers for the school with their related data
-    const teachers = await this.prisma.teacher.findMany({
-      where: {
-        user: {
-          schoolId,
-        },
-        deletedAt: null,
-      },
-      include: {
-        user: true,
-        department: true, // Direct department relationship
-        hod: {
-          include: {
-            department: true,
-          },
-        },
-        classArmSubjectTeachers: {
-          include: {
-            subject: true,
-          },
-        },
-        classArmTeachers: {
-          include: {
-            classArm: true,
-          },
-        },
-        classArmsAsTeacher: true, // Direct class teacher assignments
-      },
-    });
-
-    // Calculate statistics using real data
-    const totalTeachers = teachers.length;
-    const activeTeachers = teachers.filter((t) => t.status === TeacherStatus.ACTIVE).length;
-    const inactiveTeachers = teachers.filter((t) => t.status === TeacherStatus.INACTIVE).length;
-    const onLeaveTeachers = teachers.filter((t) => t.status === TeacherStatus.ON_LEAVE).length;
-
-    // Process teacher data
-    const teachersData = teachers.map((teacher) => {
-      // Get department name from direct relationship
-      const department = teacher.department?.name || 'Unassigned';
-
-      // Get subjects taught
-      const subjects = teacher.classArmSubjectTeachers.map((cast) => cast.subject.name);
-
-      // Get classes assigned
-      const classesAssigned = [
-        ...teacher.classArmTeachers.map((cat) => cat.classArm.name),
-        ...teacher.classArmsAsTeacher.map((classArm) => classArm.name),
-      ];
-
-      // Calculate experience (years since join date)
-      const joinDate = teacher.joinDate;
-      const experience = Math.floor(
-        (Date.now() - new Date(joinDate).getTime()) / (365.25 * 24 * 60 * 60 * 1000),
-      );
-
-      // Map employment type from enum to string
-      const employmentTypeMap = {
-        [EmploymentType.FULL_TIME]: 'full-time' as const,
-        [EmploymentType.PART_TIME]: 'part-time' as const,
-        [EmploymentType.CONTRACT]: 'contract' as const,
-      };
-      const employment = employmentTypeMap[teacher.employmentType];
-
-      // Map status from enum to string
-      const statusMap = {
-        [TeacherStatus.ACTIVE]: 'active' as const,
-        [TeacherStatus.INACTIVE]: 'inactive' as const,
-        [TeacherStatus.ON_LEAVE]: 'on-leave' as const,
-      };
-      const status = statusMap[teacher.status];
-
-      // Get last login from user
-      const lastLogin = teacher.user.lastLoginAt?.toISOString() || null;
-
-      return {
-        id: teacher.id,
-        firstName: teacher.user.firstName,
-        lastName: teacher.user.lastName,
-        email: teacher.user.email || '',
-        phone: teacher.user.phone || '',
-        department,
-        subjects,
-        employment,
-        experience,
-        qualification: teacher.qualification || 'Not specified',
-        joinDate: joinDate.toISOString().split('T')[0], // YYYY-MM-DD format
-        lastLogin,
-        status,
-        classesAssigned,
-        avatar: teacher.user.avatarUrl,
-      };
-    });
-
-    return {
-      stats: {
-        totalTeachers,
-        activeTeachers,
-        inactiveTeachers,
-        onLeaveTeachers,
-      },
-      teachers: teachersData,
-    };
+    return this.teacherService.getTeachersViewData(userId);
   }
 
   async getSingleTeacherDetails(userId: string, teacherId: string): Promise<SingleTeacherDetails> {
-    // First, get the user's school ID
-    const user = await this.prisma.user.findUnique({
-      where: { id: userId },
-      select: { schoolId: true },
-    });
-
-    if (!user?.schoolId) {
-      throw new Error('User not found or not associated with a school');
-    }
-
-    const schoolId = user.schoolId;
-
-    // Get the specific teacher with all necessary data
-    const teacher = await this.prisma.teacher.findFirst({
-      where: {
-        id: teacherId,
-        user: {
-          schoolId,
-        },
-        deletedAt: null,
-      },
-      include: {
-        user: {
-          include: {
-            address: true,
-          },
-        },
-        department: true,
-        hod: {
-          include: {
-            department: true,
-          },
-        },
-        classArmSubjectTeachers: {
-          include: {
-            subject: true,
-          },
-        },
-        classArmTeachers: {
-          include: {
-            classArm: {
-              include: {
-                students: true,
-              },
-            },
-          },
-        },
-        classArmsAsTeacher: {
-          include: {
-            students: true,
-          },
-        },
-      },
-    });
-
-    if (!teacher) {
-      throw new Error('Teacher not found or not accessible');
-    }
-
-    // Get department name
-    const department = teacher.department?.name || 'Unassigned';
-
-    // Get subjects taught
-    const subjects = teacher.classArmSubjectTeachers.map((cast) => cast.subject.name);
-
-    // Get classes assigned
-    const classesAssigned = [
-      ...teacher.classArmTeachers.map((cat) => cat.classArm.name),
-      ...teacher.classArmsAsTeacher.map((classArm) => classArm.name),
-    ];
-
-    // Calculate experience (years since join date)
-    const joinDate = teacher.joinDate;
-    const experience = Math.floor(
-      (Date.now() - new Date(joinDate).getTime()) / (365.25 * 24 * 60 * 60 * 1000),
-    );
-
-    // Map employment type from enum to string
-    const employmentTypeMap = {
-      [EmploymentType.FULL_TIME]: 'full-time' as const,
-      [EmploymentType.PART_TIME]: 'part-time' as const,
-      [EmploymentType.CONTRACT]: 'contract' as const,
-    };
-    const employment = employmentTypeMap[teacher.employmentType];
-
-    // Map status from enum to string
-    const statusMap = {
-      [TeacherStatus.ACTIVE]: 'active' as const,
-      [TeacherStatus.INACTIVE]: 'inactive' as const,
-      [TeacherStatus.ON_LEAVE]: 'on-leave' as const,
-    };
-    const status = statusMap[teacher.status];
-
-    // Get last login from user
-    const lastLogin = teacher.user.lastLoginAt?.toISOString() || null;
-
-    // Check if teacher is HOD
-    const isHOD = teacher.hod.length > 0;
-    const hodDepartment = isHOD ? teacher.hod[0].department.name : null;
-
-    // Calculate total students and average class size
-    const allStudents = [
-      ...teacher.classArmTeachers.flatMap((cat) => cat.classArm.students),
-      ...teacher.classArmsAsTeacher.flatMap((classArm) => classArm.students),
-    ];
-    const totalStudents = allStudents.length;
-    const averageClassSize =
-      classesAssigned.length > 0 ? Math.round(totalStudents / classesAssigned.length) : 0;
-
-    // Format date of birth
-    const dateOfBirth = teacher.user.dateOfBirth
-      ? teacher.user.dateOfBirth.toISOString().split('T')[0]
-      : null;
-
-    // Format address
-    const address = teacher.user.address
-      ? {
-          street1: teacher.user.address.street1,
-          street2: teacher.user.address.street2,
-          city: teacher.user.address.city,
-          state: teacher.user.address.state,
-          country: teacher.user.address.country,
-        }
-      : null;
-
-    return {
-      id: teacher.id,
-      firstName: teacher.user.firstName,
-      lastName: teacher.user.lastName,
-      email: teacher.user.email || '',
-      phone: teacher.user.phone || '',
-      department,
-      subjects,
-      employment,
-      experience,
-      qualification: teacher.qualification || 'Not specified',
-      joinDate: joinDate.toISOString().split('T')[0], // YYYY-MM-DD format
-      lastLogin,
-      status,
-      classesAssigned,
-      avatar: teacher.user.avatarUrl,
-      dateOfBirth,
-      gender: teacher.user.gender === 'MALE' ? 'Male' : 'Female',
-      stateOfOrigin: teacher.user.stateOfOrigin,
-      address,
-      isHOD,
-      hodDepartment,
-      totalStudents,
-      averageClassSize,
-    };
+    return this.teacherService.getSingleTeacherDetails(userId, teacherId);
   }
 }
